requirements.txt
asgiref==3.3.4
dj-database-url==0.5.0
Django==3.2.3
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-jsonapi==4.2.0
gunicorn==20.1.0
inflection==0.5.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
whitenoise==5.2.0
